% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{add_plus_if_needed}
\alias{add_plus_if_needed}
\title{Add "+" string if positive number}
\usage{
add_plus_if_needed(number)
}
\arguments{
\item{number}{A number}
}
\value{
An string.
}
\description{
Logic to return an "+" string if positive number.
}
\keyword{internal}
