% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bbox_perimeter}
\alias{bbox_perimeter}
\title{Bounding Box of St.Gallen perimeter map.}
\format{
An object of class \code{numeric} of length 4.
}
\usage{
bbox_perimeter
}
\description{
Bounding Box of St.Gallen perimeter map.
}
\keyword{datasets}
