% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{color_palette}
\alias{color_palette}
\title{Get color palette}
\usage{
color_palette(name)
}
\arguments{
\item{name}{name of the color palette}
}
\value{
character string of HEX codes.
}
\description{
Get St.Gallen color palettes
}
\keyword{internal}
