% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_boxes.R
\name{create_box}
\alias{create_box}
\title{Create boxes}
\usage{
create_box(input, df_box_beobachtung, df_box_referenz, indicator)
}
\arguments{
\item{input}{input}

\item{df_box_beobachtung}{df_box_beobachtung() reactive data}

\item{df_box_referenz}{df_box_referenz() reactive data}

\item{indicator}{"Indicator" variable value from both datasets}
}
\value{
an HTML string.
}
\description{
Create boxes
}
\keyword{internal}
