% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_boxes.R
\name{create_box_dur_stay}
\alias{create_box_dur_stay}
\title{Create box for "dur_stay" indicator}
\usage{
create_box_dur_stay(input, df_box_beobachtung, df_box_referenz, indicator)
}
\arguments{
\item{input}{input}

\item{df_box_beobachtung}{df_box_beobachtung() reactive data}

\item{df_box_referenz}{df_box_referenz() reactive data}

\item{indicator}{Indicator "dur_stay" value from both datasets}
}
\value{
an HTML string.
}
\description{
Create box for "dur_stay" indicator
}
\keyword{internal}
