% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_boxes.R
\name{create_box_pct}
\alias{create_box_pct}
\title{Create box for indicator with percent points}
\usage{
create_box_pct(input, df_box_beobachtung, df_box_referenz, indicator)
}
\arguments{
\item{input}{input}

\item{df_box_beobachtung}{df_box_beobachtung() reactive data}

\item{df_box_referenz}{df_box_referenz() reactive data}

\item{indicator}{Indicator "bed_occ" value from both datasets}
}
\value{
an HTML string.
}
\description{
Create box for indicator with percent points
}
\keyword{internal}
