% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_boxes.R
\name{create_box_title}
\alias{create_box_title}
\title{Create the titles of the boxes with a popover icon}
\usage{
create_box_title(title, popover_content, input, input_id)
}
\arguments{
\item{title}{title of the box}

\item{popover_content}{content of the popover}

\item{input}{input}
}
\value{
an html object.
}
\description{
Create the titles of the boxes with a popover icon
}
\keyword{internal}
