% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gt_tables.R
\name{create_gt_beobachtungsregion}
\alias{create_gt_beobachtungsregion}
\title{Create gt table beobachtungsregion}
\usage{
create_gt_beobachtungsregion(data, input)
}
\arguments{
\item{data}{data}

\item{input}{input}
}
\value{
a gt object
}
\description{
Create gt table beobachtungsregion
}
\keyword{internal}
