% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gt_tables.R
\name{create_gt_data_beobachtung}
\alias{create_gt_data_beobachtung}
\title{Create data for gt table beobachtungsregion}
\usage{
create_gt_data_beobachtung(
  df_beobachtungsregion,
  df_box_beobachtung,
  months_selected
)
}
\arguments{
\item{df_beobachtungsregion}{df_beobachtungsregion}

\item{df_box_beobachtung}{df_box_beobachtung}

\item{months_selected}{months_selected}
}
\value{
a dataframe.
}
\description{
Create data for gt table beobachtungsregion
}
\keyword{internal}
