% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gt_tables.R
\name{create_gt_data_referenzregion}
\alias{create_gt_data_referenzregion}
\title{Create data for gt table referenzregion}
\usage{
create_gt_data_referenzregion(
  df_referenzregion,
  df_box_referenz,
  months_selected,
  gt_data_beobachtung
)
}
\arguments{
\item{df_referenzregion}{df_referenzregion}

\item{df_box_referenz}{df_box_referenz}

\item{months_selected}{months_selected}

\item{gt_data_beobachtung}{gt_data_beobachtung}
}
\value{
a dataframe.
}
\description{
Create data for gt table referenzregion
}
\keyword{internal}
