% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gt_tables.R
\name{create_gt_referenzregion}
\alias{create_gt_referenzregion}
\title{Create gt table for referenzregion}
\usage{
create_gt_referenzregion(gt_table_data_referenz, gt_data_beobachtung, input)
}
\arguments{
\item{gt_table_data_referenz}{gt_table_data_referenz}

\item{gt_data_beobachtung}{gt_data_beobachtung}

\item{input}{input}
}
\value{
a gt object.
}
\description{
Create gt table for referenzregion
}
\keyword{internal}
