% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gt_tables.R
\name{create_gt_title}
\alias{create_gt_title}
\title{Create title for gt table}
\usage{
create_gt_title(input, input_region)
}
\arguments{
\item{input}{input}

\item{input_region}{input region, either input$beobachtungsjahr or input$beobachtungsjahr}
}
\value{
an html object
}
\description{
Create title for gt table
}
\keyword{internal}
