% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_months.R
\name{create_months_ui}
\alias{create_months_ui}
\title{Create Beobachtungsmonate month dropdown input UI}
\usage{
create_months_ui(data, input, ns, label_text)
}
\arguments{
\item{data}{Reactive dataset}

\item{input}{input}

\item{ns}{ns}

\item{label_text}{label of the month UI}
}
\value{
a shinyWidgets airDatepickerInput input.
}
\description{
Create the month dropdown using `shinyWidgets::airDatepickerInput()` based on
the beobachtung data and "beobachtungsjahr" input.
}
\keyword{internal}
