% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dest}
\alias{dest}
\title{Destinationen.}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
dest
}
\description{
Destinationen.
}
\keyword{datasets}
