% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dest_mun}
\alias{dest_mun}
\title{Gemeinden in den Destinationen.}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 122 rows and 3 columns.
}
\usage{
dest_mun
}
\description{
Gemeinden in den Destinationen.
}
\keyword{datasets}
