% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dest_sub}
\alias{dest_sub}
\title{Subregionen der Destinationen.}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 18 rows and 3 columns.
}
\usage{
dest_sub
}
\description{
Subregionen der Destinationen.
}
\keyword{datasets}
