% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_ueberblick}
\alias{df_ueberblick}
\title{Data used in the mod_ueberblick_tab.R}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11232 rows and 174 columns.
}
\usage{
df_ueberblick
}
\description{
A subset of official St.Gallen tourism data.
}
\keyword{datasets}
