% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{down_arrow}
\alias{down_arrow}
\title{Arrow down}
\usage{
down_arrow(color = "#8a2432")
}
\arguments{
\item{color}{Hex color of the arrow}
}
\value{
HTML arrow down.
}
\description{
Create an html arrow down
}
\keyword{internal}
