% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Filter raw data based on some UI inputs}
\usage{
filter_data(data, input_region, input)
}
\arguments{
\item{data}{dataset `sgtourism::df_ueberblick`.}

\item{input_region}{input region, either input$beobachtungsregion or
input$referenzregion.}

\item{input}{Ueberblick input beobachtungsregion, beobachtungsjahr and
referenzjahr.}
}
\value{
a filtered data.frame
}
\description{
This function create a filtered data based on the UI inputs referenzregion or
beobachtungsregion, beobachtungsjahr and referenzjahr. The function creates
a `Referenz` variable to be used in other functions in the server.
}
\keyword{internal}
