% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_boxes.R
\name{get_box_data}
\alias{get_box_data}
\title{Get data for box tables}
\usage{
get_box_data(data, months_selected, input)
}
\arguments{
\item{data}{data.frame produced by `filter_data()`.}

\item{months_selected}{reactive vector months_selected()}

\item{input}{input}
}
\value{
a data.frame to be used with `create_box_*()` functions.
}
\description{
Prepare data to be used in `create_box()`, `create_box_dur_stay()` and
`create_box_pct()`. Returns a data.frame which summarized my indicators
based on filtered data and months selected. For detailed breakdown by markt,
see `get_box_data_markt()`.
}
\keyword{internal}
