% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_country_vars}
\alias{get_country_vars}
\title{helper to get needed variable names for selected country}
\usage{
get_country_vars(country_name)
}
\arguments{
\item{country_name}{country name from interactive input}
}
\value{
variable names for arrivals and overnight stays
}
\description{
helper to get needed variable names for selected country
}
\keyword{internal}
