% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_timelines.R
\name{get_end_tl}
\alias{get_end_tl}
\title{Get end of timelines (ending at last complete year )}
\usage{
get_end_tl(yr1, yr2)
}
\arguments{
\item{yr1}{year 1}

\item{yr2}{year 2}
}
\value{
end yr for timeline plots
}
\description{
Get end of timelines (ending at last complete year )
}
\keyword{internal}
