% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_number_icon}
\alias{get_number_icon}
\title{Get arrow up or down relative to number}
\usage{
get_number_icon(number, color_positive = "#009fe3", color_negative = "#8a2432")
}
\arguments{
\item{number}{A number}

\item{color_positive}{Color of the positive number}

\item{color_negative}{Color of the negative number}
}
\value{
An arrow up or arrow down HTML string.
}
\description{
Logic to return an hex color code. If number is 0 or if not a number, returns an empty string.
}
\keyword{internal}
