% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_timelines.R
\name{get_start_tl}
\alias{get_start_tl}
\title{Get start of timelines (min length: 10 yrs)}
\usage{
get_start_tl(yr1, yr2)
}
\arguments{
\item{yr1}{year 1}

\item{yr2}{year 2}
}
\value{
start yr for timeline plots
}
\description{
Get start of timelines (min length: 10 yrs)
}
\keyword{internal}
