% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_year.R
\name{get_years}
\alias{get_years}
\title{Get years for UI}
\usage{
get_years(data)
}
\arguments{
\item{data}{dataset with an "Jahr" variable}
}
\value{
A list of vectors.
}
\description{
Get a list of year values based on the data for the UI.
}
\keyword{internal}
