% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_aggregat}
\alias{lookup_aggregat}
\title{lookup table for region names and corresponding values of var Aggregat in data set.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 48 rows and 2 columns.
}
\usage{
lookup_aggregat
}
\description{
lookup table for region names and corresponding values of var Aggregat in data set.
}
\keyword{datasets}
