% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{max_month_actual}
\alias{max_month_actual}
\title{helper to get the most recent months of the actual year}
\usage{
max_month_actual(df)
}
\arguments{
\item{df}{data frame for which the month has to be recognized}
}
\value{
Number of most recent month
}
\description{
helper to get the most recent months of the actual year
}
\keyword{internal}
