% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_countries}
\alias{meta_countries}
\title{Metadata used for country and region names.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 80 rows and 4 columns.
}
\usage{
meta_countries
}
\description{
Metadata used for country and region names.
}
\keyword{datasets}
