% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_details_tab.R
\name{mod_details_tab_server}
\alias{mod_details_tab_server}
\title{mod_details_tab_server Server Functions}
\usage{
mod_details_tab_server(id)
}
\arguments{
\item{id}{Internal parameters for shiny.}
}
\description{
mod_details_tab_server Server Functions
}
\keyword{internal}
