% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_details_tab.R
\name{mod_details_tab_ui}
\alias{mod_details_tab_ui}
\title{mod_details_tab_ui UI Function}
\usage{
mod_details_tab_ui(id)
}
\arguments{
\item{id}{Internal parameters for shiny.}
}
\description{
Shiny Module for the "details" page.
}
\keyword{internal}
