% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_perimeter_tab.R
\name{mod_perimeter_tab_server}
\alias{mod_perimeter_tab_server}
\title{perimeter_tab Server Functions}
\usage{
mod_perimeter_tab_server(id)
}
\arguments{
\item{id}{Internal parameters for shiny.}
}
\description{
perimeter_tab Server Functions
}
\keyword{internal}
