% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_perimeter_tab.R
\name{mod_perimeter_tab_ui}
\alias{mod_perimeter_tab_ui}
\title{perimeter_tab UI Function}
\usage{
mod_perimeter_tab_ui(id)
}
\arguments{
\item{id}{Internal parameters for shiny.}
}
\description{
A shiny Module.
}
\keyword{internal}
