% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ueberblick_tab.R
\name{mod_ueberblick_tab_server}
\alias{mod_ueberblick_tab_server}
\title{ueberblick_tab Server Functions}
\usage{
mod_ueberblick_tab_server(id)
}
\arguments{
\item{id}{Internal parameters for shiny.}
}
\description{
ueberblick_tab Server Functions
}
\keyword{internal}
