% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ueberblick_tab.R
\name{mod_ueberblick_tab_ui}
\alias{mod_ueberblick_tab_ui}
\title{ueberblick_tab UI Function}
\usage{
mod_ueberblick_tab_ui(id)
}
\arguments{
\item{id}{Internal parameters for shiny.}
}
\description{
A shiny Module.
}
\keyword{internal}
