% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{month_abb_de}
\alias{month_abb_de}
\title{Get German month abbreviations}
\usage{
month_abb_de()
}
\value{
An string.
}
\description{
Get a string containing the abbreviations of German months.
}
\keyword{internal}
