% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_months.R
\name{select_months}
\alias{select_months}
\title{Create numeric month vector based on monat input}
\usage{
select_months(input)
}
\arguments{
\item{input}{monat input}
}
\value{
The return value, if any, from executing the function.
}
\description{
Extract the month as numeric from the input$monat date character and returns
a vector of all months between the two months selected (if multiple selection)
or, if single selection, return a single month as number.
}
\keyword{internal}
