% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{text_with_popover_icon}
\alias{text_with_popover_icon}
\title{Text with clickable question icon with popover text}
\usage{
text_with_popover_icon(text, popover_content, input_id)
}
\arguments{
\item{text}{text to show}

\item{popover_content}{text content of the popover}

\item{input_id}{unique ID to make the popover work}
}
\value{
an text followed by a clickable icon button with a popover text.
}
\description{
Text with clickable question icon with popover text
}
\keyword{internal}
