% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{up_arrow}
\alias{up_arrow}
\title{Arrow up}
\usage{
up_arrow(color = "#009fe3")
}
\arguments{
\item{color}{Hex color of the arrow}
}
\value{
HTML arrow up.
}
\description{
Create an html arrow up
}
\keyword{internal}
